+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 * Upgraded dependencies:
     * argonaut-6.2
 * Set `secure` request attribute correctly in blaze server
-* Rename `ProcessApp.main` to `ProcessApp.run` to avoid overload confusion
+* Rename `ProcessApp.main` to `ProcessApp.process` to avoid overload confusion
 
 # v0.16.0-M1 (2017-04-08)
 * Fix type of `AuthedService.empty`